+++ rip-grep-idk/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "rip-grep-idk"
+name = "rip_grep"
 version = "0.1.0"
 edition = "2021"
